packages: continue to adjust tests and code
There is a flake in some of the tests: a file is created
in one process, but is not visible in another. What makes
no sense is that the root process does the unshare, but it is
still acting like name spaces are not correctly shared.

Go runtime makes me want to scream sometimes.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>