checkpoint commit.
Dirents when fuse are in use will have a name that is the JSON-ized Attr.

9p2000.L sure made a lot of foundational errors, one of them being
that Dirents do not contain the same info as Getattr.

This is not at all like plan 9; on plan 9, readdir returned a stream of
data and the dirents had the same info that a stat would return.

The result was that on plan 9, unlike linux, a readdir gave you all you need.
You NEVER saw a readdir follow by lots of stats, which you do on Linux/Unix/Darwin/etc.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>