client: add an os the Cmd and WithServerGOOS Option
In order to support remote systems like freebsd, we
will need to know that remote OS.

Sadly, the SSH version strings are an utter mess, and very
few of them (save cpud) cleanly show the remote kernel.

Add WithServerGOOS so that programs (such as cpu) can use it.

The client package can use this variable when setting, e.g,
NFS mount options.

Because we did not anticipate the need for this, all code
which uses the os variable will need to deal with the zero
value; they should assume os == "" is the same as
os == "linux"

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>